Added abstracted drawing functions

+++ TEXT.C
@@ -8,14 +8,14 @@ void setTextPosition(int x, int y)
 	printf("%c[%d;%df", 0x1B, y+1, x+1);
 }
 
-void setTextColor(int fg, int bg)
+void setTextColor(byte fg, byte bg)
 {
 	// Escape code to set fore/background colours
 	printf("%c[1;%d;%dm", 0x1B, fg, bg);
 }
 
 // Write text to a specific position with foreground and background colours
-void putText(char* string, int fg, int bg, int x, int y)
+void putText(char* string, byte fg, byte bg, int x, int y)
 {
 	setTextPosition(x, y);
 	setTextColor(fg, bg);
